+++ goty-client/src/components/submission/Giveaway.tsx
@@ -5,7 +5,7 @@ import { generateRules } from '../../util/generate-rules'
 import { useSelector, useStore } from 'react-redux'
 import { selectProperties } from '../../state/properties/selectors'
 import { createUpdateEnteredGiveawayAction } from '../../state/submission/actions'
-import { GotyRadioSet, RadioOption } from '../restyled/GotyRadioButton'
+import { GotyRadioSet, RadioOption } from '../styled-controls/GotyRadioButton'
 
 export interface GiveawayProps {
   readonly: boolean
